+++ scripts/tours.js
@@ -1,33 +1,39 @@
 import { MODULE_ID } from "./const.js";
+import { localize } from "./helpers.js";
 
 export function registerMyTours() {
-  game.tours.register(MODULE_ID, "welcome", new Tour(welcomeTour));
+  const tourData = getTourData();
+  game.tours.register(MODULE_ID, "welcome", new SidebarTour(tourData.welcome));
 }
 
-const welcomeTour = {
-  title: `${MODULE_ID}.tours.welcome.title`,
-  description: `${MODULE_ID}.tours.welcome.desc`,
-  canBeResumed: true,
-  display: true,
-  steps: [
-    {
-      id: `compendiums`,
-      selector: `.tabs>a[data-tab='compendium']`,
-      title: `${MODULE_ID}.tours.welcome.compendiums.title`,
-      content: `${MODULE_ID}.tours.welcome.compendiums.content`,
-      sidebarTab: `compendium`,
+function getTourData() {
+  return {
+    welcome: {
+      title: localize(`tours.welcome.title`),
+      description: localize(`tours.welcome.desc`),
+      canBeResumed: true,
+      display: true,
+      steps: [
+        {
+          id: `compendiums`,
+          selector: `.tabs>a[data-tab='compendium']`,
+          title: localize(`tours.welcome.steps.compendiums.title`),
+          content: localize(`tours.welcome.steps.compendiums.content`),
+          sidebarTab: "compendium",
+        },
+        {
+          id: `adventures`,
+          selector: `input[type="search"]`,
+          title: localize(`tours.welcome.steps.adventures.title`),
+          content: localize(`tours.welcome.steps.adventures.content`),
+        },
+        {
+          id: `maps`,
+          selector: `input[type="search"]`,
+          title: localize(`tours.welcome.steps.maps.title`),
+          content: localize(`tours.welcome.steps.maps.content`),
+        },
+      ],
     },
-    {
-      id: `adventures`,
-      selector: `.compendium[data-entry-id='fist-of-the-ruby-phoenix-addons.fist-of-the-ruby-phoenix-addons']`,
-      title: `${MODULE_ID}.tours.welcome.adventures.title`,
-      content: `${MODULE_ID}.tours.welcome.adventures.content`,
-    },
-    {
-      id: `maps`,
-      selector: `.compendium[data-entry-id='fist-of-the-ruby-phoenix-addons.fist-of-the-ruby-phoenix-addons-maps']`,
-      title: `${MODULE_ID}.tours.welcome.maps.title`,
-      content: `${MODULE_ID}.tours.welcome.maps.content`,
-    },
-  ],
-};
+  };
+}